fix some static analyzer warnings

+++ WinPort/src/ConsoleOutput.cpp
@@ -271,7 +271,7 @@ void ConsoleOutput::ScrollOutputOnOverflow(SMALL_RECT &area)
 	if ( (height - _scroll_region.top) < 2 || width==0)
 		return;
 
-	_temp_chars.resize(width * (height - 1) );
+	_temp_chars.resize(size_t(width) * (height - 1) );
 	if (_temp_chars.empty())
 		return;
